y2023/src/test/kotlin/adventofcode/y2023/Day14Test.kt
package adventofcode.y2023

import org.junit.jupiter.api.Assertions.assertEquals
import org.junit.jupiter.api.Test

internal class Day14Test {

    @Test
    fun solvePartOne_a() {
        val input = """
            O....#....
            O.OO#....#
            .....##...
            OO.#O....O
            .O.....O#.
            O.#..O.#.#
            ..O..#O..O
            .......O..
            #....###..
            #OO..#....""".trimIndent()
        assertEquals(136, Day14.solvePartOne(input))
    }


    @Test
    fun solvePartTwo() {
        val input = """
            O....#....
            O.OO#....#
            .....##...
            OO.#O....O
            .O.....O#.
            O.#..O.#.#
            ..O..#O..O
            .......O..
            #....###..
            #OO..#....""".trimIndent()
        assertEquals(64, Day14.solvePartTwo(input))
    }
}



